+++ src/DependencyInjection/CompilerPass/ReplaceFailoverAdaptersLocatorCompilerPass.php
@@ -1,0 +1,43 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Webf\Flysystem\DsnBundle\DependencyInjection\CompilerPass;
+
+use Symfony\Component\DependencyInjection\Argument\IteratorArgument;
+use Symfony\Component\DependencyInjection\Argument\TaggedIteratorArgument;
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Webf\Flysystem\DsnBundle\DependencyInjection\WebfFlysystemDsnExtension;
+use Webf\Flysystem\DsnBundle\Flysystem\FailoverAdaptersLocator;
+use Webf\FlysystemFailoverBundle\DependencyInjection\WebfFlysystemFailoverExtension;
+
+class ReplaceFailoverAdaptersLocatorCompilerPass implements CompilerPassInterface
+{
+    public function process(ContainerBuilder $container): void
+    {
+        $serviceId = WebfFlysystemFailoverExtension::FAILOVER_ADAPTERS_LOCATOR_SERVICE_ID;
+        $tagName = WebfFlysystemDsnExtension::ADAPTER_TAG_NAME;
+
+        if ($container->hasDefinition($serviceId)) {
+            $definition = $container->getDefinition($serviceId);
+
+            /** @var IteratorArgument $failoverAdapters */
+            $failoverAdapters = $definition->getArgument(0);
+
+            foreach ($failoverAdapters->getValues() as $reference) {
+                $container
+                    ->getDefinition($reference->__toString())
+                    ->addTag($tagName)
+                ;
+            }
+
+            $definition
+                ->setClass(FailoverAdaptersLocator::class)
+                ->setArguments([
+                    new TaggedIteratorArgument($tagName),
+                ])
+            ;
+        }
+    }
+}