+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### ✨ New features
 
   * Add `ServiceAdapterFactory`
+  * Register `FailoverAdapterFactory` as service
 
 ## v0.1.0 (August 31, 2021)